Create jar for GCStandalone
